Fix Colab install so it works on current (2026) Colab
The pinned Dec-2024 stack no longer installed on current Colab, and the
setup cell swallowed the failure (check=False) -- notebooks then ran on
Colab's stock torch/transformers and died later with cryptic errors
(bitsandbytes>=..., torchvision::nms, unexpected 'dtype' kwarg).

- Setup cell installs with uv, not pip: uv's resolver installs the pinned
  stack cleanly where pip's resolver aborts the whole install.
- requirements-colab.txt: pin torchvision==0.20.1 / torchaudio==2.5.1 to
  match torch 2.5.1 (fixes "operator torchvision::nms does not exist").
- Code used the transformers-5.x `dtype=` kwarg against the transformers
  4.46 pin; fixed dtype= -> torch_dtype= in the 3 AutoModelForCausalLM
  .from_pretrained calls (local_llm.py, trainers.py x2). unsloth/vllm
  dtype= left alone.
- Rebuild all 9 notebooks; 133 tests pass, NB0 offline dry-run clean.
- Add WORKSHOP_GUIDE.md: minimal participant runbook (open -> T4 -> Run all).

+++ llm_utils/local_llm.py
@@ -118,7 +118,7 @@ class LocalLM:
             ) if self.load_in_4bit else None
             self.tokenizer = AutoTokenizer.from_pretrained(mid)
             self.model = AutoModelForCausalLM.from_pretrained(
-                mid, quantization_config=quant, dtype=dtype,
+                mid, quantization_config=quant, torch_dtype=dtype,
                 device_map="auto", attn_implementation="sdpa",  # FA2 needs sm_80+
             )
             self.model.eval()
